+++ R/mapedit.R
@@ -1,0 +1,20 @@
+#' mapedit: interactive editing and selection for geospatial data
+#'
+#' mapedit, a RConsortium funded project, provides interactive
+#' tools to incorporate in geospatial workflows that require editing or selection
+#' of spatial data.
+#'
+#' @section Edit:
+#' \itemize{
+#' \item{\code{\link{editMap}}}
+#' \item{\code{\link{editFeatures}}}
+#' \item{Shiny edit module \code{\link{editModUI}}, \code{\link{editMod}}}
+#' }
+#'
+#' #' @section Edit:
+#' \itemize{
+#' \item{\code{\link{selectMap}}}
+#' \item{\code{\link{selectFeatures}}}
+#' \item{Shiny edit module \code{\link{selectModUI}}, \code{\link{selectMod}}}
+#' }
+"_PACKAGE"