+++ R/query.R
@@ -1,21 +1,27 @@
 #' Interactively Select Map Features
 #'
-#' @param x map to use
-#'
+#' @param x \code{leaflet} or \code{mapview} map to use for selection
 #' @param ... other arguments
-#'
 #' @example ./inst/examples/examples_select.R
 #' @export
 selectMap <- function(x, ...) {
   UseMethod("selectMap")
 }
 
+#' @name selectMap
+#' @param styleFalse,styleTrue names \code{list} of CSS styles used
+#'          for selected (\code{styleTrue}) and deselected (\code{styleFalse})
+#' @param ns \code{string} name for the Shiny \code{namespace} to use.  The \code{ns}
+#'          is unlikely to require a change.
+#' @param viewer \code{function} for the viewer.  See Shiny \code{\link[shiny]{viewer}}.
 #' @export
 selectMap.leaflet <- function(
   x = NULL,
   styleFalse = list(fillOpacity = 0.2, weight = 1, opacity = 0.4),
   styleTrue = list(fillOpacity = 0.7, weight = 3, opacity = 0.7),
-  ns = "mapedit-select"
+  ns = "mapedit-select",
+  viewer = shiny::paneViewer(),
+  ...
 ) {
   stopifnot(!is.null(x), inherits(x, "leaflet"))
 
@@ -27,7 +33,10 @@ selectMap.leaflet <- function(
   )
 
   ui <- miniUI::miniPage(
-    miniUI::miniContentPanel(selectModUI(ns), height=NULL, width=NULL),
+    miniUI::miniContentPanel(
+      selectModUI(id = ns, height = "97%"),
+      height=NULL, width=NULL
+    ),
     miniUI::gadgetTitleBar("Select Features on Map", right = miniUI::miniTitleBarButton("done", "Done", primary = TRUE))
   )
 
@@ -55,6 +64,7 @@ selectMap.leaflet <- function(
   shiny::runGadget(
     ui,
     server,
+    viewer = viewer,
     stopOnCancel = FALSE
   )
 }